birdtag-backend/src/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
cos-python-sdk-v5==1.9.23
python-dotenv==1.0.0